requirements.txt
Flask==3.1.2
face_recognition==1.3.0
opencv-python==4.9.0.80
imutils==0.5.4
pillow==11.3.0
openai==1.54.4
python-dotenv==1.0.1
dlib==20.0.0
numpy==2.3.2
werkzeug==3.1.3
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1
deepface==0.0.79
mtcnn==0.1.1
tf-keras
